🔥 Tidy up accountsstorages + remove unnecessary methids for FuzzClient
