+++ examples/fuzz-tests/simple-cpi-6/trident-tests/fuzz_tests/fuzz_0/fuzz_instructions.rs
@@ -134,5 +134,5 @@ impl IxOps for InitializeCaller {
 #[derive(Default)]
 pub struct FuzzAccounts {
     _program: AccountsStorage<ProgramStore>,
-    signer_caller: AccountsStorage<Keypair>,
+    signer_caller: AccountsStorage<KeypairStore>,
 }
